Use automatic route code splitting
