+++ frontend/src/routes/clients.$id.tsx
@@ -1,11 +1,15 @@
-// Copyright 2024 New Vector Ltd.
+// Copyright 2024, 2025 New Vector Ltd.
 // Copyright 2024 The Matrix.org Foundation C.I.C.
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 // Please see LICENSE in the repository root for full details.
 
+import { useSuspenseQuery } from "@tanstack/react-query";
 import { queryOptions } from "@tanstack/react-query";
+import { notFound } from "@tanstack/react-router";
 import { createFileRoute } from "@tanstack/react-router";
+import OAuth2ClientDetail from "../components/Client/OAuth2ClientDetail";
+import Layout from "../components/Layout";
 import { graphql } from "../gql";
 import { graphqlRequest } from "../graphql";
 
@@ -17,7 +21,7 @@ const QUERY = graphql(/* GraphQL */ `
   }
 `);
 
-export const query = (id: string) =>
+const query = (id: string) =>
   queryOptions({
     queryKey: ["oauth2Client", id],
     queryFn: ({ signal }) =>
@@ -27,4 +31,19 @@ export const query = (id: string) =>
 export const Route = createFileRoute("/clients/$id")({
   loader: ({ context, params }) =>
     context.queryClient.ensureQueryData(query(params.id)),
+  component: ClientDetail,
 });
+
+function ClientDetail(): React.ReactElement {
+  const { id } = Route.useParams();
+  const {
+    data: { oauth2Client },
+  } = useSuspenseQuery(query(id));
+  if (!oauth2Client) throw notFound();
+
+  return (
+    <Layout>
+      <OAuth2ClientDetail client={oauth2Client} />
+    </Layout>
+  );
+}
